+++ engine/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.5.4
 ### Changes
 * Added an explanatory comment as to why all the directories that the engine server creates inside the engine data directory must be created with `0777` permissions